lastnews_project/requirements.txt
Django==3.1
django-filter==2.3.0
djangorestframework==3.11.1
celery==4.4.7
psycopg2-binary==2.8.5
requests==2.24.0
django-extensions==3.0.5
xmltodict==0.12.0
